+++ client/package.json
@@ -21,6 +21,9 @@
   },
   "dependencies": {
     "@hidden-identity/server": "workspace:*",
+    "@tanstack/react-query": "^4.35.3",
+    "@trpc/client": "^10.38.3",
+    "@trpc/react-query": "^10.38.3",
     "autoprefixer": "^10.4.15",
     "axios": "^1.5.0",
     "classnames": "^2.3.2",
