+++ client/vite.config.ts
@@ -21,6 +21,11 @@ export default defineConfig((env) => ({
         rewrite: (path) => path.replace(/^\/api/, ""),
         ws: true,
       },
+      "/trpc": {
+        target: "http://localhost:3001",
+        changeOrigin: true,
+        rewrite: (path) => path.replace(/^\/trpc/, ""),
+      },
     },
   },
   define: {
