+++ server/package.json
@@ -12,6 +12,9 @@
   "exports": {
     ".": {
       "types": "./declarations/types/index.d.ts"
+    },
+    "./server": {
+      "types": "./declarations/index.d.ts"
     }
   },
   "scripts": {
@@ -48,7 +51,9 @@
     "typescript": "*"
   },
   "dependencies": {
+    "@trpc/server": "^10.38.3",
     "@types/body-parser": "^1.19.2",
-    "body-parser": "^1.20.2"
+    "body-parser": "^1.20.2",
+    "zod": "^3.22.2"
   }
 }
