+++ server/src/routes/game.ts
@@ -1,136 +1,178 @@
-import { type Application } from 'express-ws'
 import {
   addGame,
+  getGame,
+  retrieveGame,
+  updateStatus,
+} from '../database/gameDB/base.ts'
+import {
+  assignPlayerToRole,
+} from '../database/gameDB/seating.ts'
+import {
   addPlayerStatus,
   addPlayer,
   assignRoles,
   clearPlayerStatus,
-  getGame,
   kickPlayer,
-  retrieveGame,
   setPlayerFate,
   setPlayerNote,
   setPlayerOrder,
   toggleDeadvote,
-  updateStatus,
-  assignPlayerToRole,
   setAlignment,
-} from '../database/gameDB.ts'
+} from '../database/gameDB/player.ts'
 import { setupTestGames } from '../testGames.ts'
+import { gmProcedure, playerProcedure, publicProcedure } from '../trpc.ts'
+import { z } from 'zod'
+import { alignmentShape } from '../types/Role.ts'
+import { poisonStatusShape, drunkStatusShape, customStatusShape, gameStatusShape } from '../types/UnifiedGame.ts'
+import { gameIdShape, playerAndGameIdShape, roleShape } from './baseApiShapes.ts'
 
-export function useGame (app: Application): void {
-  setupTestGames()
-  app.post('/game', (req, res) => {
-    const { hash: gameId, oldGameId } = req.body
+setupTestGames()
 
-    if (oldGameId) {
-      console.log(`recieved old gameID: ${oldGameId}, updating old game`)
+export const gameRoutes = {
+  getGame: publicProcedure
+    .input(gameIdShape)
+    .query(async ({ input: { gameId } }) => {
+      // Retrieve users from a datasource, this is an imaginary database
+      const game = getGame(gameId)
 
-      const oldGame = retrieveGame(oldGameId)
-      oldGame.update({
-        ...oldGame.readOnce(),
-        nextGameId: gameId,
-      })
-    }
-    console.log(`creating new game ${gameId}`)
-    addGame(gameId)
-    const game = getGame(gameId)
-    console.log(`Game created ${gameId}`)
-    console.log(`responding with: ${JSON.stringify(game)}`)
-    res.send(game)
-  })
+      return game
+    }),
+  createGame: publicProcedure
+    .input(
+      z.intersection(
+        gameIdShape,
+        z.object({
+          oldGameId: z.string().optional(),
+        })),
+    )
+    .mutation(async ({ input: { gameId, oldGameId } }) => {
+      addGame(gameId)
+      if (oldGameId) {
+        console.log(`recieved old gameID: ${oldGameId}, updating old game`)
 
-  app.get('/game/:gameId', (_req, res) => {
-    const gameId = _req.params.gameId
-    const game = getGame(gameId)
-    if (!game) {
-      res.status(404)
-    }
-    res.send(game)
-  })
+        const oldGame = retrieveGame(oldGameId)
+        oldGame.update({
+          ...oldGame.readOnce(),
+          nextGameId: gameId,
+        })
+      }
 
-  app.post('/add_player', (req, res) => {
-    const { player, gameId } = req.body
-    addPlayer(gameId, player)
-    res.status(200)
-    res.send({
-      player,
-    })
-  })
-
-  app.post('/kick_player', (req, res) => {
-    const { player, gameId } = req.body
-    kickPlayer(gameId, player)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/order_player', (req, res) => {
-    const { player, rightNeighbor, gameId } = req.body
-    setPlayerOrder(gameId, player, rightNeighbor)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/assign_roles', (req, res) => {
-    const { roles, gameId } = req.body
-    assignRoles(gameId, roles)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/decide_fate', (req, res) => {
-    const { player, gameId, dead } = req.body
-    setPlayerFate(gameId, player, dead)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/add_status_effect', (req, res) => {
-    const { player, gameId, playerStatus } = req.body
-    addPlayerStatus(gameId, player, playerStatus)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/clear_status_effect', (req, res) => {
-    const { player, gameId, playerStatusId } = req.body
-    clearPlayerStatus(gameId, player, playerStatusId)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/set_player_note', (req, res) => {
-    const { player, gameId, note } = req.body
-    setPlayerNote(gameId, player, note)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/dead_vote', (req, res) => {
-    const { player, gameId, voteUsed } = req.body
-    toggleDeadvote(gameId, player, voteUsed)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/manual_status', (req, res) => {
-    const { gameId, status } = req.body
-    updateStatus(gameId, status)
-    res.status(200)
-    res.send({})
-  })
-
-  app.post('/assign_role', (req, res) => {
-    const { gameId, player, role } = req.body
-    assignPlayerToRole(gameId, player, role)
-    res.status(200)
-    res.send({})
-  })
-  app.post('/set_alignment', (req, res) => {
-    const { gameId, player, alignment } = req.body
-    setAlignment(gameId, player, alignment)
-    res.status(200)
-    res.send({})
-  })
+      return getGame(gameId)
+    },
+    ),
+  addPlayer: playerProcedure
+    .input(playerAndGameIdShape)
+    .mutation(({ input: { gameId, player } }) => {
+      addPlayer(gameId, player)
+      return player
+    }),
+  kickPlayer: gmProcedure
+    .input(playerAndGameIdShape)
+    .mutation(({ input: { gameId, player } }) => {
+      kickPlayer(gameId, player)
+      return null
+    }),
+  setPlayerOrder: playerProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({ rightNeighbor: z.string().nullable() }),
+      ),
+    ).mutation(({ input: { gameId, player, rightNeighbor } }) => {
+      setPlayerOrder(gameId, player, rightNeighbor)
+      return null
+    }),
+  assignRoles: gmProcedure
+    .input(
+      z.intersection(
+        gameIdShape,
+        z.object({ roles: z.array(roleShape) }),
+      ),
+    )
+    .mutation(({ input: { gameId, roles } }) => {
+      assignRoles(gameId, roles)
+    }),
+  decideFate: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({ dead: z.boolean() }),
+      ),
+    )
+    .mutation(({ input: { gameId, player, dead } }) => {
+      setPlayerFate(gameId, player, dead)
+    }),
+  addPlayerStatus: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({
+          playerStatus:
+          z.union([poisonStatusShape, drunkStatusShape, customStatusShape]),
+        }),
+      ),
+    )
+    .mutation(({ input: { gameId, player, playerStatus } }) => {
+      addPlayerStatus(gameId, player, playerStatus)
+    }),
+  clearPlayerStatus: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({
+          playerStatusId: z.string(),
+        }),
+      ),
+    )
+    .mutation(({ input: { gameId, player, playerStatusId } }) => {
+      clearPlayerStatus(gameId, player, playerStatusId)
+    }),
+  setPlayerNote: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({
+          note: z.string(),
+        }),
+      ),
+    )
+    .mutation(({ input: { gameId, player, note } }) => {
+      setPlayerNote(gameId, player, note)
+    }),
+  setDeadVote: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({ voteUsed: z.boolean() }),
+      ),
+    ).mutation(({ input: { gameId, player, voteUsed } }) => {
+      toggleDeadvote(gameId, player, voteUsed)
+    }),
+  setGameStatus: gmProcedure
+    .input(
+      z.intersection(
+        gameIdShape,
+        z.object({ status: gameStatusShape }),
+      ),
+    ).mutation(({ input: { gameId, status } }) => {
+      updateStatus(gameId, status)
+    }),
+  assignRole: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({ role: roleShape }),
+      ),
+    ).mutation(({ input: { gameId, player, role } }) => {
+      assignPlayerToRole(gameId, player, role)
+    }),
+  setAlignment: gmProcedure
+    .input(
+      z.intersection(
+        playerAndGameIdShape,
+        z.object({ alignment: alignmentShape }),
+      ),
+    ).mutation(({ input: { gameId, player, alignment } }) => {
+      setAlignment(gameId, player, alignment)
+    }),
 }
