+++ server/src/routes/script.ts
@@ -1,11 +1,17 @@
-import { type Application } from 'express-ws'
 import { setScript } from '../database/scriptDB.ts'
+import { gmProcedure } from '../trpc.ts'
+import { gameIdShape, roleShape } from './baseApiShapes.ts'
+import { z } from 'zod'
 
-export function useScript (app: Application): void {
-  app.post('/set_script', (req, res) => {
-    const { gameId, script } = req.body
-    setScript(gameId, script)
-    res.status(200)
-    res.send({})
-  })
+export const scriptRoutes = {
+  setScript: gmProcedure
+    .input(
+      z.intersection(
+        gameIdShape,
+        z.object({ script: z.array(z.object({ id: roleShape })) }),
+      ))
+    .mutation(async ({ input: { gameId, script } }) => {
+      // Retrieve users from a datasource, this is an imaginary database
+      setScript(gameId, script)
+    }),
 }
